+++ ucas-os/kernel/syscall/syscall.c
@@ -6,7 +6,7 @@ long (*syscall[NUM_SYSCALLS])();
 void handle_syscall(regs_context_t *regs, uint64_t interrupt, uint64_t cause)
 {
     regs->sepc = regs->sepc + 4;
-    if (syscall[regs->a7] == handle_miss)
+    if ((size_t)syscall[regs->a7] == (size_t)handle_miss)
         handle_miss(regs, interrupt, cause);
     // debug_info(regs, interrupt, cause);
     regs->a0 = syscall[regs->a7](regs->a0,
